Updates model.router to support multiple model file in the API
