+++ src/server/schema/model-file.schema.ts
@@ -6,6 +6,7 @@ export const modelFileSchema = z.object({
   url: z.string().url().min(1, 'You must select a file'),
   sizeKB: z.number(),
   type: z.nativeEnum(ModelFileType),
+  primary: z.boolean().default(false),
 });
 
-export type ModelFileProps = z.infer<typeof modelFileSchema>;
+export type ModelFileInput = z.infer<typeof modelFileSchema>;
