+++ src/server/selectors/model.selector.ts
@@ -148,6 +148,7 @@ export const modelWithDetailsSelect = Prisma.validator<Prisma.ModelSelect>()({
           virusScanMessage: true,
           scannedAt: true,
           rawScanResult: true,
+          primary: true,
         },
       },
       // runStrategies: {
