+++ src/utils/file-helpers.ts
@@ -1,4 +1,5 @@
-import { ModelFileFormat } from '@prisma/client';
+import { ModelFileFormat, ModelFileType, Prisma, ScanResultCode } from '@prisma/client';
+import { ModelFileInput } from '~/server/schema/model-file.schema';
 
 export function getModelFileFormat(filename: string) {
   if (filename.endsWith('.safetensors')) return ModelFileFormat.SafeTensor;
@@ -6,4 +7,25 @@ export function getModelFileFormat(filename: string) {
     return ModelFileFormat.PickleTensor;
 
   return null;
+}
+
+const unscannedFile = {
+  scannedAt: null,
+  scanRequestedAt: null,
+  rawScanResult: Prisma.JsonNull,
+  virusScanMessage: null,
+  virusScanResult: ScanResultCode.Pending,
+  pickleScanMessage: null,
+  pickleScanResult: ScanResultCode.Pending,
+};
+
+export function prepareFile(file: ModelFileInput) {
+  return {
+    ...file,
+    ...unscannedFile,
+    format:
+      file.type === ModelFileType.Model || file.type === ModelFileType.PrunedModel
+        ? getModelFileFormat(file.name)
+        : null,
+  };
 }
